extract text-color method to composable
